Authorization over getTaskById, updateTaskStatus and deleteTaskById

+++ src/tasks/tasks.service.ts
@@ -15,13 +15,19 @@ export class TasksService {
         private taskRepository: TaskRepository
     ) {}
 
-    async getTasks(filterDto: GetTaskFilterDto): Promise<Task[]> {
-        return this.taskRepository.getTasks(filterDto);
+    async getTasks(
+        filterDto: GetTaskFilterDto,
+        user: User,
+    ): Promise<Task[]> {
+        return this.taskRepository.getTasks(filterDto, user);
     }
 
     
-    async getTaskById(id:number): Promise<Task> {
-        const found = await this.taskRepository.findOne(id);
+    async getTaskById(
+        id:number,
+        user: User,
+    ): Promise<Task> {
+        const found = await this.taskRepository.findOne({ where: { id, userId: user.id }});
         if(!found){
             throw new NotFoundException(`Task with ID '${id}' not found`);
         }
@@ -37,8 +43,11 @@ export class TasksService {
     }
     
 
-    async  deleteTaskById(id: number): Promise<DeleteResult> {
-        const result = await this.taskRepository.delete(id);
+    async  deleteTaskById(
+        id: number,
+        user: User,
+    ): Promise<DeleteResult> {
+        const result = await this.taskRepository.delete({id, userId: user.id});
         if(result.affected === 0){
             throw new NotFoundException(`Task with ID '${id}' not found`);
         }
@@ -46,8 +55,12 @@ export class TasksService {
         return result;
     }
 
-    async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-        const task = await this.getTaskById(id);
+    async updateTaskStatus(
+        id: number, 
+        status: TaskStatus,
+        user: User,
+    ): Promise<Task> {
+        const task = await this.getTaskById(id, user);
         task.status = status;
         await task.save()
